Refactored PPP package directory structure by adding 'definitions', and including constants and errors to be in said directory. Added 'abstractions' to 'definitions'. Minor refactoring in argument building logic of CLI.
