+++ pmix/ppp/odkchoices.py
@@ -1,5 +1,5 @@
 """Module for the OdkChoices class."""
-from pmix.ppp.error import InvalidLanguageException
+from pmix.ppp.definitions.error import InvalidLanguageException
 
 
 class OdkChoices:
